app/requirements.txt
descartes==1.1.0
diameter-clustering==0.0.1
gast==0.3.3
geog==0.0.2
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
gevent-websocket==0.10.1
glob2==0.7
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
haversine==2.5.1
HeapDict==1.0.1
imagesize==1.2.0
matplotlib-venn==0.11.6
numpy==1.19.2
numpy-groupies==0.9.14
pandas==1.3.2
pathlib2==2.3.5
pathtools==0.1.2
plotly==4.14.3
reverse-geocoder==1.5.1
simplekml==1.3.6
streamlit==1.7.0